breaking: rename elements_hist to hist_elemental_prevalence in pymatviz/histograms.py
